Move `TestContext` struct and initialization method to `test_context_test.go`
